waste classification/requirements.txt
tensorflow
tensorflow-hub
tensorflow-model-optimization
numpy
matplotlib
scikit-learn
pillow
tqdm
